+++ NEWS.md
@@ -1,7 +1,8 @@
 # fauxnaif 0.6.0
 
 * No longer produce a warning when no replacements are made.
-* Deprecate scoped_na_if() functions and suggest replacing them with `dplyr::across()`.
+* Deprecate scoped na_if() functions and suggest replacing them with `dplyr::across()`.
+* Deprecate `fauxnaif::na_if()` and suggest replacing it with `na_if_in()`.
 
 # fauxnaif 0.5.6